split sick_tim3xx_common into common and usb specific stuff
