+++ CMakeLists.txt
@@ -33,10 +33,10 @@ gencfg()
 #target_link_libraries(${PROJECT_NAME} another_library)
 #rosbuild_add_boost_directories()
 #rosbuild_link_boost(${PROJECT_NAME} thread)
-rosbuild_add_executable(sick_tim310s01 src/sick_tim310s01.cpp src/sick_tim3xx_common.cpp src/sick_tim310s01_parser.cpp src/abstract_parser.cpp)
-rosbuild_add_executable(sick_tim310 src/sick_tim310.cpp src/sick_tim3xx_common.cpp src/sick_tim310_parser.cpp src/abstract_parser.cpp)
-rosbuild_add_executable(sick_tim310_1130000m01 src/sick_tim310_1130000m01.cpp src/sick_tim3xx_common.cpp src/sick_tim310_1130000m01_parser.cpp src/abstract_parser.cpp)
-rosbuild_add_executable(sick_tim551_2050001 src/sick_tim551_2050001.cpp src/sick_tim3xx_common.cpp src/sick_tim551_2050001_parser.cpp src/abstract_parser.cpp)
+rosbuild_add_executable(sick_tim310s01 src/sick_tim310s01.cpp src/sick_tim3xx_common.cpp src/sick_tim3xx_common_usb.cpp src/sick_tim310s01_parser.cpp src/abstract_parser.cpp)
+rosbuild_add_executable(sick_tim310 src/sick_tim310.cpp src/sick_tim3xx_common.cpp src/sick_tim3xx_common_usb.cpp src/sick_tim310_parser.cpp src/abstract_parser.cpp)
+rosbuild_add_executable(sick_tim310_1130000m01 src/sick_tim310_1130000m01.cpp src/sick_tim3xx_common.cpp src/sick_tim3xx_common_usb.cpp src/sick_tim310_1130000m01_parser.cpp src/abstract_parser.cpp)
+rosbuild_add_executable(sick_tim551_2050001 src/sick_tim551_2050001.cpp src/sick_tim3xx_common.cpp src/sick_tim3xx_common_usb.cpp src/sick_tim551_2050001_parser.cpp src/abstract_parser.cpp)
 rosbuild_add_executable(sick_tim3xx_datagram_test test/sick_tim3xx_datagram_test.cpp src/sick_tim310s01_parser.cpp src/abstract_parser.cpp)
 
 target_link_libraries(sick_tim310s01 usb-1.0)
