+++ include/sick_tim3xx/sick_tim3xx_common_usb.h
@@ -1,10 +1,10 @@
 /*
  * Copyright (C) 2013, Osnabrück University
  * All rights reserved.
- *
+ * 
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- *
+ * 
  *     * Redistributions of source code must retain the above copyright
  *       notice, this list of conditions and the following disclaimer.
  *     * Redistributions in binary form must reproduce the above copyright
@@ -13,7 +13,7 @@
  *     * Neither the name of Osnabrück University nor the names of its
  *       contributors may be used to endorse or promote products derived from
  *       this software without specific prior written permission.
- *
+ * 
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -36,22 +36,56 @@
  *
  */
 
-#include <sick_tim3xx/sick_tim3xx_common_usb.h>
-#include <sick_tim3xx/sick_tim310s01_parser.h>
+#ifndef SICK_TIM3XX_COMMON_USB_H_
+#define SICK_TIM3XX_COMMON_USB_H_
 
-int main(int argc, char **argv)
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <libusb-1.0/libusb.h>
+
+#include <sick_tim3xx/sick_tim3xx_common.h>
+
+namespace sick_tim3xx
 {
-  ros::init(argc, argv, "sick_tim310s01");
 
-  sick_tim3xx::SickTim310S01Parser* parser = new sick_tim3xx::SickTim310S01Parser();
-  sick_tim3xx::SickTim3xxCommonUsb s((sick_tim3xx::AbstractParser*)parser);
+class SickTim3xxCommonUsb : public SickTim3xxCommon
+{
+public:
+  SickTim3xxCommonUsb(AbstractParser* parser);
+  virtual ~SickTim3xxCommonUsb();
 
-  int result = s.init();
-  while (ros::ok() && (result == EXIT_SUCCESS))
-  {
-    ros::spinOnce();
-    result = s.loopOnce();
-  }
+protected:
+  virtual int init_device();
+  virtual int close_device();
 
-  return result;
-}
+  /// Send a SOPAS command to the device and print out the response to the console.
+  virtual int sendSOPASCommand(const char* request);
+
+  /// Read a datagram from the device.
+  /**
+   * \param [in] receiveBuffer data buffer to fill
+   * \param [in] bufferSize max data size to write to buffer (result should be 0 terminated)
+   * \param [out] actual_length the actual amount of data written
+   */
+  virtual int get_datagram(unsigned char* receiveBuffer, int bufferSize, int* actual_length);
+
+private:
+  static const unsigned int USB_TIMEOUT = 500; // milliseconds
+
+  ssize_t getSOPASDeviceList(libusb_context *ctx, uint16_t vendorID, uint16_t productID, libusb_device ***list);
+  void freeSOPASDeviceList(libusb_device **list);
+
+  void printUSBDeviceDetails(struct libusb_device_descriptor desc);
+  void printUSBInterfaceDetails(libusb_device* device);
+  void printSOPASDeviceInformation(ssize_t numberOfDevices, libusb_device** devices);
+
+  // libusb
+  libusb_context *ctx_;
+  ssize_t numberOfDevices_;
+  libusb_device **devices_;
+  libusb_device_handle *device_handle_;
+};
+
+} /* namespace sick_tim3xx */
+#endif /* SICK_TIM3XX_COMMON_USB_H_ */
