+++ src/sick_tim310_1130000m01.cpp
@@ -36,7 +36,7 @@
  *
  */
 
-#include <sick_tim3xx/sick_tim3xx_common.h>
+#include <sick_tim3xx/sick_tim3xx_common_usb.h>
 #include <sick_tim3xx/sick_tim310_1130000m01_parser.h>
 
 int main(int argc, char **argv)
@@ -44,9 +44,9 @@ int main(int argc, char **argv)
   ros::init(argc, argv, "sick_tim310_1130000m01");
 
   sick_tim3xx::SickTim3101130000M01Parser* parser = new sick_tim3xx::SickTim3101130000M01Parser();
-  sick_tim3xx::SickTim3xxCommon s((sick_tim3xx::AbstractParser*)parser);
+  sick_tim3xx::SickTim3xxCommonUsb s((sick_tim3xx::AbstractParser*)parser);
 
-  int result = s.init_usb();
+  int result = s.init();
   while (ros::ok() && (result == EXIT_SUCCESS))
   {
     ros::spinOnce();
